+++ clients/pkg/promtail/targets/manager.go
@@ -1,6 +1,8 @@
 package targets
 
 import (
+	"fmt"
+
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/pkg/errors"
@@ -80,7 +82,7 @@ func NewTargetManagers(
 			targetScrapeConfigs[WindowsEventsConfigs] = append(targetScrapeConfigs[WindowsEventsConfigs], cfg)
 
 		default:
-			return nil, errors.New("unknown scrape config")
+			return nil, fmt.Errorf("no valid target scrape config defined for %q", cfg.JobName)
 		}
 	}